Switch server create -f to using open-uri lib.
Doing so means you can pass a URL as the -f argument and it'll read the contents of that URL in instead of only reading in a file from the local filesystem. Annoyingly open-uri returns a StringIO object for a remote URL, and a File object for local files, which have differing methods to get the size of the contents.

Existing behaviour isn't broken doing this.