Use negatable switch for base64 encoding
This changes the behaviour of the options a bit which isn't
backwards compatible, but it's a less used option so I think it's ok.

 --no_base64 becomes --no-base64 which is easy enough

but -e now means *do* encode base64 (and is default), which
inverts the old behaviour. Bad form to change this but again, a rarely
used option and the new behaviour is more consistent.