Uses a config fixture file to enable testing
A config file located in the ambiguous `spec/brightbox/config` was
required to enable the specs to run against the VCR recordings.

This uses a new fixture in `spec/fixtures/config_files/config` to enable
anyone to run the tests without knowing the undocumented steps required.

The rarer case, using a live connection to record traffic, is not yet
covered.