Tests and refactors setting connection behaviour
Using fog, when using the compute service (confusingly also called
the connection) you should never need to be change the service a
resource is from.

This behaviour is deprecated after fog 0.8 so using a later version
begins outputting deprecation warnings when using the CLI.

This tests the methods in Brightbox::Account that were attempting this
behaviour. It appears to be a redundant operation.

Removes one use of Account#conn_account

After removing the attempt to set the service, #conn_account is the same
as @conn#account so removing to simply things.