Move and renamed initial specs
This moves the specs we do have to be part of the unit tests with better
isolation between them and their VCR records.

This has forced us to patch the recordings so they can run independently

Mocha expectations are interferring causing problems where
Brightbox::Api.find is still a mock from an earlier spec.