Removes need for config at class level
By including the name of the config file in the description, a brittle
ordering dependency was set up where a global config was required before
the main module was loaded.

This forced the BBConfig.config &block syntax upon us for very little
benefit and created some really odd situations after refectoring.

This removes the reference to the $config, loads the main GLI module
once and handles the config inside GLI's pre/post blocks in the method
call not the require context.

This should fix some future bugs and allow us to narrow the config to
not be global.