Use globals rather than constants for IO
Normally STDOUT == $stdout so they can be used interchangably. Trying to
test the captured output and it is inconsistent because we are logging
to either STDOUT or $stdout (or the err version).

This makes it impossible to capture ALL output and check since the IO is
a mixture.

We can't cleanly reset the constants so using the globals.