Adds Firewall policy subcommand
Can now update a firewall policy's name or description

Tests to confirm the update has occured (by making independent API calls)
fail out of sequence due to global config and token caching. They may or
may not work depending on which specs have run before so are pending.

The specs are triggering an issue in account specs as well.