Removes custom cassette naming scheme
The technique to automatically name a VCR cassette based on the spec is
available in VCR itself so this switches to that.

The built in functionality handles names better so they have been
renamed to match.

The cassettes have been moved to spec/cassettes to save a few
characters in the names as well.