Locks nokogiri to the last 1.8.7 version
Nokogiri 1.6 drops Ruby 1.8.7 support and declares specifically in the
gemspec. We don't even USE nokogiri but it is a painful dependency of
fog.

So we declare the version we can work with.

We also nudge up to fog 1.14 which is identical except for a looser
nokogiri dependency that breaks bundling on 1.8.7 without the lock
above.