Refactors BBConfig internals
This break up some of the combined methods to cut down on side effects
and give a bit more control over the internals.

BBConfig#dir would create the directory if it was missing and we could
not access the path without grabbing the instance variable.

This has been replaced by #config_directory which returns the path.

BBConfig#config still behaves as before, creating the directory if
missing and loading the file if available.