Adds test constants for both client versions
Either an API client or a User Application can be used to access the API
and this is now available in the specs.

Use either:
  config = Brightbox::BBConfig.new(:directory => API_CLIENT_CONFIG)
  config = Brightbox::BBConfig.new(:directory => USER_APP_CONFIG)

Unfortunately there is still a global config ($config) still in use so
certain specs/code relying on the global will need to be mocked at some
level.