Cleans up BBConfig#finish
BBConfig#finish appears to have been buried under side effects. The
config was being saved in the #save_refresh_token and in other places
the same #save! method was being called in commands.

The #finish method should run after each command because of the post
block so we were saving the config twice for the config commands.

The inline code for saving the access token has been extracted to
the #save_access_token method.

Both access and refresh tokens should be being saved to disk as files,
so the references to the config can be replaced in future work.