Extracts default client out as side effect
A side effect of #client_name is raising an error if a default config
was not specified in the config.

For the only use in the `config` command we can make it safer because we
do not want the side effect (raising an error) just getting the value
from the config if available.