Sets all specs into tmp $HOME dirs
Not very efficient but more correct for the moment, create a temporary
directory for each spec and use it for the $HOME.

This means that the user running the specs should NEVER have their home
directory interfered with.

This also helps highlight specs that fail on Travis CI because they have
been making do with user's config.