Refactors token management inside config
This brings some logic from error handling and isolates the token
management in the configuration.

The BBConfig class gets refactored around the token handling.

The unauthenticated logic from ErrorParser is removed. In trying to recover from
errors globally we are again going down the route of relying on a global config
to tell it to renew tokens.

Fixes token renewal when adding a new "user" client

Changes mean a lot of VCR recording needed updating to account for the
new authentication flows in place.