Updates client tests to not rely on particular keys
From the OAuth side of things it doesn't matter which type of client you
are authorising but the grant type matters.

From the CLI perspective the client_id and secret can be the same
because the original implementation switched based on the name of the
key in the config.

That was brittle if you added both to one "client" which is possible
because the keys differ.

This is a first step in unifying the format of sections again.