Fixes find_or_set_default_account when unauthorised
Through a chain of usage #account is called during the authorisation
steps which then tries to make API calls when we know the calls will
fail.

Really we should only do this helper call when adding the client the
first time. This is creating extra API traffic and only helps one case.

As an alternative we should guard against making the call if the acces
token has been voided. At the moment that is still hit/miss.