Removes need for app_id key
The config FILE can contain many SECTIONS. Each section is independent
and contains one client's details.

Possibly for detecting if the config was for an old API client or a new
User App, two keys were used for exactly the same thing the client_id
and secret passed to fog.

Being different they could be placed side by side in one config section
merging two clients with random behaviour.

This removes the "app" version of the keys which will break existing
configuration setup using a prerelease version of the CLI.