Renames username value from email
"email" is inaccurate since it can be a username like usr-12345 or an
email address. The OAuth option is again username.

This also slims down the 'options' passed to #fetch_refresh_tokens since
none are optional for it to work based on current implementation.
Sending blank values to fog will just trigger API errors.