Don't prompt to rerun command when added new user
The code to renew tokens included a prompt to retry the command being
aborted. The same code was used for adding a new user.

The moves the prompt so when adding a user for the first time there is
no prompt telling you to retry your command (which was adding the user)