Allows adding of duplicate user accounts
The user add command assumed the user application identifier was going
to be unique and used it as the config section name.

Adding a second user with the same application then caused an error but
it was decided this was a valid use case.

This adds code to add a numerical suffix in the form (_\d+) if the app
identifier is already in the config.

Previously the user had to try the command again with an alias of their
own whilst now we select one for them (that they can manage and change
as before).