Fixes default account code
The config itself requests a list of accounts to select the default rather than
relying on the Api connection. This Api connection is broken when adding a new
config.

We use the config details as given to do a standard fog call to get the listing.

Unfortunately this appears to create a large amount of VCR churn