Removes example configs from test setup
A number of issues have been allowed to lurk by the presence of config files for
all specs. A number of specs have then required us to go in and remove the
config to test things.

This no longer sets up the configs so each spec runs in it's own HOME and it is
up to the spec to create the config required.