Makes Config#account safer
What a palaver! So #account called #find_or_set_default_account which queries
the API. The Api connection needed a fog configuration which may need to
reference #account...

The ONLY times you would ever need this is:

* Added a new client and you want to check
* The user is trying a command with no default and no argument

We do the first as a sensible optimisation and we guard around the second.

Now we have a safer accessor.