Updates #client_named? to support old alias
In the refactoring we replaced support for the old alias with the
scanning the section headers.

Section headers are supposed to be interchangable with alias now
but there will probably be cases where this is not true. Need a
better API to manage the config file / clients than messing with
hashes.