Remove hard dependency on JSON gem
fog uses MultiJson so should be running the fastest available JSON
parser in the environment.

By isolating our one use of JSON we lose the need for the hard
dependency.

Unless you are using Bundler to isolate a command all we have done is
added a native dependency to installing the gem.