Replace Fog version constant
A mistake in `fog-core` removed `Fog::VERSION` in a minor update which
meant the HTTP headers broke.

This uses the replacement `Fog::Core::VERSION` for the moment.

There was an issue with that originally which was why the original was
left in place but hopefully this update will be the end of the matter.