Support encrypted stored passwords
When a password is needed for an account, check to see if it has been
gpg-encrypted and stored in a local file first, and attempt to decrypt
it with gpg.

It looks for ~/.brightbox/john@johnleach.co.uk.password.gpg and just
executes "gpg --decrypt" on it. The idea is that you'll combine it
with a gpg key daemon, such as GNOME's seahorse.

If the gpg binary isn't in the path, decryption will not be attempted
and you'll just be prompted for your password as usual.

You can encrypt a password by running this command:

gpg --encrypt --recipient john@brightbox.co.uk > ~/.brightbox/john@johnleach.co.uk.password.gpg

and then typing your password, followed by CTRL+D

Any carriage returns in the decrypted password are ignored. You can
"forget" the password by just removing that file.