Test login action passing URL flags
This tests `api-url` and `auth-url` are stored in the config when passed
as arguments.

There is the potential for a bug in that the VCR test recordings rewrite
the URLs against a dev endpoint which is also the default one. So the
test as stands may allow a regression where the default is always used.

Hacking the tests demonstrates correct behaviour (with failures of the
tests) but it could be broken later.