[#105296840] Default to embedded app if not set
This adds a default user application set of credentials (which still
MUST be used with a username/password) which are used if the ID/secret
are not configured.

This will allow us to cut down the complexity of setting up new
configurations since only the email (username) and password (prompt)
will be needed.