Rename section name
There has been a confusing merge of "config" and "client" which is
getting in the way of development.

There is a `BBConfig` object that wraps A LOT of code and an Ini file.

The Ini config file consists of several "sections" which represent a set
of client credentials (aka client).

Name change is a first step at clearing things up.