Simplify `#client_name` method and `#config`
This makes a number of accessors simpler by moving the decision making
up from to `#new` as well and creating a missing config dir.

Was implemented as a side effect of calling `config` but that meant
other methods were forced for call config to create/load the Ini file.

This ended up with situations where `default_client` could not be read
from the object without ensure it had called `config` to side effect its
loading.