Refactor out `$config` global
This replaces `$config` with a singleton class holding a reference to a
`Brightbox::BBConfig`

This is only slightly better since it allows specs to be stubbed per
example without risking screwing up the global variable affecting ALL
specs.

It does place the files in the preferred directory given the namespace.