Clean up spec for nonexistent config directory
This fixes a recurring issue where running specs in random order can
create a named config directory ahead of a spec testing if does not
exist.

This now creates a temp directory and immediately discards it before
testing for it.

Any directories created by spec errors will at least exist in tmp and
should not clash.