Fix caching attempts when directory missing
Running specs in a random order could trigger a number of failures
related to naughty API classes attempting to cache IDs whilst the
directory used did not exist.

This alters the `#cache_path` method to be read only without side effect
and creates it when  `#cache_id` is called and it does not exist.