Spec stability changes for `#config_directory_exists?`
Still seeing an error (rarely) in random runs. Digging a bit more and
calling `config` in RSpec's `let` created it _after_ the temp directory
was created. This initialized the new class and that creates the
directory again.

This calls `config` earlier, then deletes the directory before testing
the method.

There is still a very rare case when the directory may/may not exist for
deleting.

May be a race condition in FileUtils but unable to observe the error
amongst a batch of 10 random sets.