Correctly set client name for `login`
`#client_alias` method was used by mistake instead of `config_alias`
variable.

This meant the config was set to the default account whilst clearing the
access token.

The `#renew_tokens` call used the correct client since it is passed in
as an argument. However the presence of the refresh token was the reason
`#renew_tokens` would hit the branch without using the password.

This would forced a password prompt and a confusing message.