Fix issue if `maintenance_weekday` not in JSON
Whilst highly unlikely, this fixes a raised TypeError if the response
from the API does not have the `maintenance_weekday` field in the
response.

The most likely case of this is in specs where only a subset of the
response is needed for other testing concerns. This bug would require
those to have this field, just to avoid this bug.