Add `--cloud-ip` argument to new server requests
This allows the creation of servers that immediately map a Cloud IP upon
being built.

The argument can accept either "true" which requests the allocation of a
new Cloud IP for the built.

Alternatively it can accept an identifier (`cli-12345`) for an unmapped
Cloud IP and that nominated IP will be reserved and mapped instead.