Update and limit `fog-brightbox` dependency
The next release of `fog-brightbox` is planned to be a major release and
drops support for Ruby 1.9 and 2.0.

It also fixes a long standing namespacing issue that is incompatible
with the tests in this codebase.

So this updates `fog-brightbox` to fix a few issues and adds a
constraint so it will not pick up a v1.0 gem if updating itself.