Fix image filtering
When https://github.com/brightbox/brightbox-cli/pull/83 was merged, the
`account` was changed from a fog model (responding to `#id`) to a hash
with `#[]` key access.

At some point, the key type has changed - potentially caused by a Ruby
or `fog` change - from a symbol to a string.

The key change meant that the image listings had a filter that was used
by default but would filter anything without an owner ID of `nil`. That
meant everything.

The workaround was using `--show-all` which disabled this filter but
then included other images that were not interesting.

This fixes the key type to `String` for now which seems to work with
several versions of Ruby and the latest fog gems.

For: #GH-97