Update `fog-brightbox` to v0.16.1
And issue in `fog-core` exposed an ordering bug where services were
being declared before autoloading was. The core change meant a large
deprecation warning and also an error.

Updating to `fog-brightbox v0.16.1` fixes the warning but it is safest
to limit fog-core to < 2.0 when deprecation behaviours were being added.

Due to upstream requirements, the dependency can not be declared in
fog-brightbox itself.

This also returns support for Ruby 1.9 which fog (core) drops in 2.0