Remove VCR identifier sanitisation
This has the unfortunate side-effect of creating conflicting entries in
VCR recordings where one run could get a 201 response, then another
request for `srv-12345` gets a 404 and the later entry took precidence.

All the identifiers used in testing are from a development instance so
are not leaking real identifiers.

When all the VCR tests correctly setup and tear down resources and can
re-record then the re-use of 12345 will just cause problems.