Remove VCR from login specs
It is nearly impossible to automate the variations of user logins at the
server side, especially when considering 2FA one-time passwords. As such
recording VCR output is low value as we can not record easily and won't
be able to automate re-records.

The new `AuthenticationHelpers` have been used to stub out most of the
traffic with some options to support 2FA accounts.