Remove limit to RSpec error output
When VCR/webmock are preventing a HTTP call, they output a very long
trace of the request that is blocked. However this is reported via the
CLI output and RSpec is trimming that down.

This removes the output formatting length so we can actually see the
entire error and the important details as part of it.