Refactor password discovery
We had two blocks of near identical code that checked an input, or a
pipeline of other methods to recover the user's password from encrypted
stores or a helper application before prompting.

Besides the potential starting input which could have been supplied,
the only other difference was a notice about tokens having expired
(explaining why the prompt was asking again).

This extracts the common code, with options out to `PasswordHelper`
along with the prompting.