Prevent API calls when login has failed
The error handling only covers failing to renewing token, which then
immediately proceeds to attempt to find a default account. This will
also fail due to invalid tokens.

This moves the call so that isn't attempted as the guaranteed failure is
a waste of time. This results in the two useless API calls were see
during failures disappearing.