Add command spec covering invalid login
This adds an additional test case for when a login attempt fails due to
the user's credentials being rejected by the server.

Not the traffic is slightly flawed as a fallback call as a client always
occurs and fails, then a lookup of accounts (to select the default) also
fails as they are not authenticated.