Handle 2FA client connections fully
This enables the new option in `fog-brightbox v1.6.0` which will raise a
different error if a user's authentication fails due to them having Two
Factor Authentication (2FA) enabled but not supplying a One Time
Password (OTP).

This results in prompts for a user's OTP during certain actions such as
the `login` command or when tokens have expired and are being renewed.

This replaces the functionality enabled by the `two_factor` setting in
configurations. 2FA enrolment is automatically discovered or ignored.