Add `min-ram` argument to Images
This allows passing an `min-ram` argument that can be used to mark an
image as requiring a minimum amount of RAM so creating a server from too
small a type will report an error.

Rather than an instance building and failing at some point as it runs
out of memory at some point.