Update Jenkinsfile with comparable Ruby versions
We primarily use GitHub actions to confirm Ruby versions work but we do
have an internal CI service double checking. However there has been some
false negatives here due to changes in v4.0 and the docker images we
have used.

This removes the older versions and adds the newer versions to be
consistent there with the supported versions.