Fix empty server groups during `servers create`
The API does not accept empty arrays of groups as each server must be
assigned to a group. A server side workaround has been in place to
account for this long standing bug where the unset `server-groups` flag
would report an empty array.

The workaround strips out empty lists and uses the default group which
is the expected behaviour. The newer API documentation had to be
reported incorrectly to account for the workaround.