Fix `volumes --delete-with-server update` switch
This had an ugly workaround involving a non-boolean default to prevent
GLI inserting a default of `false` when the option was not passed by the
user. This default meant we couldn't determine if it was set or disabled
by the user in the API call or added by the GLI framework.

The workaround made the help incorrect as it reported a "enabled"
default which is wrong as no default should be used.

This adds the `ignore_default` option so that it is non longer added
when performing an update and we can filter it out and not pass it on
to the API call.

Not filtering would result in a rename overwriting the setting for
`delete_with_server` that exists without the user wanting that.