Fix intrusive CLI switch defaults
The GLI frameworks assumes that if a switch is defined, it should always
be present with a `true` or `false`. This prevents using switches to
represent boolean API settings since defining one injects it into the
parsed arguments.

This can not be detected within CLI actions, we can not tell if the user
passed `--feature`, `--no-feature`, or did not include the flag at all.
This may result in a `update --name New` call also setting the feature
to a default of false.

The default can be abused with a non-boolean value and used to remove it
from the parameters passed to the API but being `truthy` results in the
CLI help being incorrect. It lists a default when there isn't one.

This monkey patches GLI to support a `ignore_default` option which can
be set to `true` and it is skipped when defaults are being generated.

This can then be correctly checked for, as the key is missing within the
action and we can correctly NOT send the parameter to an API call.