Remove `source` option from `images register`
The `source` option was tied to the FTP based workflow which is no
longer supported. Thus the option would trigger an unresolvable error
for the CLI user.

Using the `url` option against a publicly accessible image via HTTP(S),
such as a temp URL in Orbit is the preferred approach to register.