Move Image `arch` to own column/remove from name
Originally the image architecture was embedded into the name for display
but that caused issues with parsing, especially if parentheses we used
in the name.

It also resulted in duplication in the `images show` command as the
`arch` field was also included but the name without was not.

This splits them apart rather than combining them.