Fix `lba` `buffer_size` arguments
This appears to been broken for a while with no attempt to correct the
API requests being sent for the `buffer_size` argument.

A missing argument would result in a 0 which fails validation at the
API, or the validation would fail as a String form was being sent when
passed.