Rename and spec `#formatted_acme_domains`
Since `acme_domains` is a underlying attribute, using the same name as
the heavily formatted version to display is going to cause problems
later so this renames it to be clearer.

After trying to stub the HTTP approach was thwarted by VCR issues, this
tests using doubles and recreating the underlying fog model.