Wrap `Api#attributes` with indifferent key access
A number of bugs have been introduced or missed due to the way
attributes are exposed by fog. Fog transforms the top level of
attributes from the JSON sources (String) to Symbols, but does not act
recursively.

This results in accessing nested structures with a mix of Symbol and
String keys which inevitably is missed or incorrectly specced out.

This adds a simple wrapper around `Api#attributes` that takes the fog
attributes and allows indifferent access with Symbols or Strings.